Basic test for picorv32
This is just a basic run on PICORV a 32 bit processor with testbench where we dump the hex files, which further taken by the memory inside PICORV processor.
Still the structure of the processor needed to be explored.....
The C program will take the value from its interface to the assembly language source file and generate the hex file through RISCV GCC complier.